requirements.txt
flask>=3.0.0
yt-dlp>=2024.1.0
pywebview>=5.0
pyinstaller>=6.0
